+++ schema-wizard-core/src/main/java/org/schemawizard/core/migration/factory/impl/PostgreSqlColumnTypeFactory.java
@@ -1,7 +1,6 @@
 package org.schemawizard.core.migration.factory.impl;
 
 import org.schemawizard.core.di.annotation.Qualifier;
-import org.schemawizard.core.metadata.DatabaseProvider;
 import org.schemawizard.core.migration.factory.ColumnTypeFactory;
 import org.schemawizard.core.migration.metadata.ColumnTypeFactoryQualifier;
 import org.schemawizard.core.migration.metadata.PlainColumnType;
@@ -12,8 +11,8 @@ public class PostgreSqlColumnTypeFactory implements ColumnTypeFactory {
     public String getNative(String type) {
         if (PlainColumnType.TEXT.equalsIgnoreCase(type)) {
             return "VARCHAR";
-        } else if (PlainColumnType.DOUBLE.equalsIgnoreCase(type)) {
-            return "DOUBLE";
+        } else if (PlainColumnType.DECIMAL.equalsIgnoreCase(type)) {
+            return "NUMERIC";
         } else if (PlainColumnType.INTEGER.equalsIgnoreCase(type)) {
             return "INTEGER";
         } else if (PlainColumnType.BOOLEAN.equalsIgnoreCase(type)) {
