+++ schema-wizard-core/src/main/java/org/schemawizard/core/migration/metadata/PlainColumnType.java
@@ -3,7 +3,7 @@ package org.schemawizard.core.migration.metadata;
 public class PlainColumnType {
     public static final String INTEGER = "integer";
     public static final String BOOLEAN = "boolean";
-    public static final String DOUBLE = "double";
+    public static final String DECIMAL = "double";
     public static final String TEXT = "text";
 
     private PlainColumnType() {
